Shared error module + payload decode with byte-order parameter (#62)
