+++ src/dlt/mod.rs
@@ -1,5 +1,6 @@
 pub mod v1;
 pub mod v2;
+pub mod error;
 pub mod intern;
 pub mod payload;
 pub mod protocol;
