+++ src/dlt/v2/payload.rs
@@ -1,240 +1,22 @@
-use crate::dlt::payload::*;
+use crate::dlt::payload;
 use crate::dlt::protocol::{CNTI_CONTROL, CNTI_NON_VERBOSE, CNTI_VERBOSE};
 
 /// Decode a payload slice into a human-readable string based on content type.
+///
+/// v2 always uses big-endian byte order for payload fields.
 pub fn decode_payload(cnti: u8, raw: &[u8]) -> String {
     match cnti {
-        CNTI_VERBOSE => decode_verbose(raw),
-        CNTI_NON_VERBOSE => decode_non_verbose(raw),
-        CNTI_CONTROL => decode_control(raw),
-        _ => format!("[unknown cnti={cnti}] {}", hex_dump(raw)),
+        CNTI_VERBOSE => payload::decode_verbose(raw, true),
+        CNTI_NON_VERBOSE => payload::decode_non_verbose(raw, true),
+        CNTI_CONTROL => payload::decode_control(raw, true),
+        _ => format!("[unknown cnti={cnti}] {}", payload::hex_dump(raw)),
     }
-}
-
-fn decode_verbose(mut data: &[u8]) -> String {
-    let mut parts: Vec<String> = Vec::new();
-
-    while data.len() >= 4 {
-        let type_info = u32::from_be_bytes(data[0..4].try_into().unwrap());
-        data = &data[4..];
-
-        // Skip variable name if VARI is set
-        if type_info & DLT_TYPE_INFO_VARI != 0 {
-            if data.len() < 2 {
-                break;
-            }
-            let name_len = u16::from_be_bytes(data[0..2].try_into().unwrap()) as usize;
-            data = &data[2..];
-            if data.len() < name_len {
-                break;
-            }
-            data = &data[name_len..];
-
-            // For numeric types with VARI, there's also a unit name
-            if type_info & (DLT_TYPE_INFO_SINT | DLT_TYPE_INFO_UINT | DLT_TYPE_INFO_FLOA) != 0 {
-                if data.len() < 2 {
-                    break;
-                }
-                let unit_len = u16::from_be_bytes(data[0..2].try_into().unwrap()) as usize;
-                data = &data[2..];
-                if data.len() < unit_len {
-                    break;
-                }
-                data = &data[unit_len..];
-            }
-        }
-
-        let tyle = type_info & DLT_TYPE_INFO_TYLE;
-
-        if type_info & DLT_TYPE_INFO_BOOL != 0 {
-            if data.is_empty() {
-                break;
-            }
-            let val = data[0] != 0;
-            parts.push(if val { "true".into() } else { "false".into() });
-            data = &data[1..];
-        } else if type_info & DLT_TYPE_INFO_SINT != 0 {
-            match tyle {
-                DLT_TYLE_8BIT => {
-                    if data.is_empty() {
-                        break;
-                    }
-                    parts.push(format!("{}", data[0] as i8));
-                    data = &data[1..];
-                }
-                DLT_TYLE_16BIT => {
-                    if data.len() < 2 {
-                        break;
-                    }
-                    let v = i16::from_be_bytes(data[0..2].try_into().unwrap());
-                    parts.push(format!("{v}"));
-                    data = &data[2..];
-                }
-                DLT_TYLE_32BIT => {
-                    if data.len() < 4 {
-                        break;
-                    }
-                    let v = i32::from_be_bytes(data[0..4].try_into().unwrap());
-                    parts.push(format!("{v}"));
-                    data = &data[4..];
-                }
-                DLT_TYLE_64BIT => {
-                    if data.len() < 8 {
-                        break;
-                    }
-                    let v = i64::from_be_bytes(data[0..8].try_into().unwrap());
-                    parts.push(format!("{v}"));
-                    data = &data[8..];
-                }
-                _ => break,
-            }
-        } else if type_info & DLT_TYPE_INFO_UINT != 0 {
-            match tyle {
-                DLT_TYLE_8BIT => {
-                    if data.is_empty() {
-                        break;
-                    }
-                    parts.push(format!("{}", data[0]));
-                    data = &data[1..];
-                }
-                DLT_TYLE_16BIT => {
-                    if data.len() < 2 {
-                        break;
-                    }
-                    let v = u16::from_be_bytes(data[0..2].try_into().unwrap());
-                    parts.push(format!("{v}"));
-                    data = &data[2..];
-                }
-                DLT_TYLE_32BIT => {
-                    if data.len() < 4 {
-                        break;
-                    }
-                    let v = u32::from_be_bytes(data[0..4].try_into().unwrap());
-                    parts.push(format!("{v}"));
-                    data = &data[4..];
-                }
-                DLT_TYLE_64BIT => {
-                    if data.len() < 8 {
-                        break;
-                    }
-                    let v = u64::from_be_bytes(data[0..8].try_into().unwrap());
-                    parts.push(format!("{v}"));
-                    data = &data[8..];
-                }
-                _ => break,
-            }
-        } else if type_info & DLT_TYPE_INFO_FLOA != 0 {
-            match tyle {
-                DLT_TYLE_32BIT => {
-                    if data.len() < 4 {
-                        break;
-                    }
-                    let v = f32::from_be_bytes(data[0..4].try_into().unwrap());
-                    parts.push(format!("{v}"));
-                    data = &data[4..];
-                }
-                DLT_TYLE_64BIT => {
-                    if data.len() < 8 {
-                        break;
-                    }
-                    let v = f64::from_be_bytes(data[0..8].try_into().unwrap());
-                    parts.push(format!("{v}"));
-                    data = &data[8..];
-                }
-                _ => break,
-            }
-        } else if type_info & DLT_TYPE_INFO_STRG != 0 {
-            if data.len() < 2 {
-                break;
-            }
-            let str_len = u16::from_be_bytes(data[0..2].try_into().unwrap()) as usize;
-            data = &data[2..];
-            if data.len() < str_len {
-                break;
-            }
-            let str_data = &data[..str_len];
-            // Strip null terminator if present
-            let s = if str_data.last() == Some(&0) {
-                &str_data[..str_len - 1]
-            } else {
-                str_data
-            };
-            let decoded = String::from_utf8_lossy(s).into_owned();
-            parts.push(decoded);
-            data = &data[str_len..];
-        } else if type_info & DLT_TYPE_INFO_RAWD != 0 {
-            if data.len() < 2 {
-                break;
-            }
-            let raw_len = u16::from_be_bytes(data[0..2].try_into().unwrap()) as usize;
-            data = &data[2..];
-            if data.len() < raw_len {
-                break;
-            }
-            parts.push(hex_dump(&data[..raw_len]));
-            data = &data[raw_len..];
-        } else {
-            // Unknown type — stop decoding
-            break;
-        }
-    }
-
-    parts.join(" ")
-}
-
-fn decode_non_verbose(data: &[u8]) -> String {
-    if data.len() < 4 {
-        return format!("[non-verbose] {}", hex_dump(data));
-    }
-    let msg_id = u32::from_be_bytes(data[0..4].try_into().unwrap());
-    let rest = &data[4..];
-    format!("[non-verbose 0x{msg_id:08x}] {}", hex_dump(rest))
-}
-
-fn decode_control(data: &[u8]) -> String {
-    if data.len() < 4 {
-        return format!("[control] {}", hex_dump(data));
-    }
-    let service_id = u32::from_be_bytes(data[0..4].try_into().unwrap());
-    let service_name = if (service_id as usize) < SERVICE_ID_NAME.len() {
-        let name = SERVICE_ID_NAME[service_id as usize];
-        if name.is_empty() {
-            format!("0x{service_id:04x}")
-        } else {
-            name.to_string()
-        }
-    } else {
-        format!("0x{service_id:04x}")
-    };
-
-    let rest = &data[4..];
-    if rest.is_empty() {
-        return format!("[control {service_name}]");
-    }
-
-    // First byte of response is the return type
-    let return_type_id = rest[0] as usize;
-    let return_text = if return_type_id < RETURN_TYPE.len() {
-        let rt = RETURN_TYPE[return_type_id];
-        if rt.is_empty() {
-            hex_dump(rest)
-        } else {
-            rt.to_string()
-        }
-    } else {
-        hex_dump(rest)
-    };
-
-    format!("[control {service_name}] {return_text}")
-}
-
-fn hex_dump(data: &[u8]) -> String {
-    data.iter().map(|b| format!("{b:02x}")).collect::<Vec<_>>().join(" ")
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::dlt::payload::*;
 
     // Helper to build a verbose TypeInfo + data
     fn make_bool(val: bool) -> Vec<u8> {
